SQL Practice page
It offers multiple database servers and databases which associated with the selected server platform in the dropdown list. The user can view the database diagram and can create their own databases to practice on. It also contains a text editor which can accept the input as well as a Grid View to manipulate the query results.